requirements.txt
aiogram[redis]==3.10.0

django==4.2.7
djangorestframework==3.14.0

python-dotenv==1.0.1

redis==5.0.7

celery==5.3.4

groq==0.9.0
google-genai>=0.3.0

python-docx==1.1.2

markdown==3.6
beautifulsoup4==4.12.3

gunicorn==21.2.0

asgiref==3.7.2
sqlparse==0.4.4
pytz==2023.3

whitenoise==6.6.0